ADD navigation component
